# server/internal/api/swagger/ui/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File4Base API Engine",
    "description": "Enterprise-grade REST and Dynamic Relational DDL Database Engine for File4Base. Supports runtime schema migrations, table occurrences, relationships graph, dynamic Find Mode formula parsing, security ACL, and solution archiving. Every endpoint except the health probes, sign-in and the database selector requires a session token: sign in with POST /api/v1/auth/login and send the returned token as 'Authorization: Bearer <token>'. A session is bound to the database it signed in to.",
    "version": "0.5.0",
    "contact": {
      "name": "File4Base Core Team",
      "url": "https://github.com/file4base/file4base-app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local API Service (Direct)"
    },
    {
      "url": "http://localhost:3000",
      "description": "WebDirect Nginx Reverse Proxy"
    },
    {
      "url": "/",
      "description": "Current Host"
    }
  ],
...
  "paths": {
...
    "/api/v1/databases": {
      "get": {
        "tags": [
          "Solutions & Backup"
        ],
        "summary": "List solution databases (public: used by the sign-in screen)",
        "security": [],
        "responses": {
          "200": {
            "description": "Databases on the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Solutions & Backup"
        ],
        "summary": "Create a database and its first owner account",
        "description": "Public while ALLOW_PUBLIC_DATABASE_CREATION=true (default); otherwise requires an owner session. Never modifies an existing database.",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Database created"
          },
          "409": {
            "description": "A database with that name already exists"
          },
          "422": {
            "description": "Missing name or owner credentials"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/databases/{name}": {
      "delete": {
        "tags": [
          "Solutions & Backup"
        ],
        "summary": "Drop a database",
        "description": "Requires an owner session on that database, or the credentials of one of its owners in the request body.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Database dropped"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Database not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details for HTTP APIs",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Invalid column format"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/schemas/tables"
          }
        }
      },
      "CreateDatabaseRequest": {
        "type": "object",
        "required": [
          "database",
          "user",
          "password"
        ],
        "properties": {
          "database": {
            "type": "string",
            "example": "invoices_db"
          },
          "user": {
            "type": "string",
            "description": "Username of the first owner account"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Password of the first owner account"
          }
        }
      },
      "DatabaseList": {
        "type": "object",
        "properties": {
          "databases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "string",
            "description": "Database of the caller's session; empty when signed out"
          }
        }
      },
      "OwnerCredentials": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token returned by POST /api/v1/auth/login. Bound to one user and one database; expires after a period of inactivity (SESSION_TTL, default 12h)."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid or expired session token",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session's role or layout permissions do not allow this operation",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}